bomtool: initial commit with working CSV read/write
This new bomtool implementation allows simple rules in a master CSV
file to determine what changes to make when generating
variant-specific BOMs.  So far, seems to work well.

This is usable now for filtering CSV BOMs; next steps are to
implement Eagle read/write support so that

(1) master BOM can be read/written into Eagle files so that the data
stays with the correct parts during edits, renames, etc.

(1) filtered BOMs can be written into Eagle files so that a file is
specific to a particular variant, ready for upload to Circuithub